// src/app/modules/docente/models/docentePerfilModel.ts
export interface DocentePerfilModel {
    id_usuario: number;
    nombre_usuario: string;
    contrasena_usuario: string;
    rol_usuario: number;
    id_docente: number;
    nombres_docente: string;
    apellidos_docente: string;
    cedula_docente: string;
    fechaNacimiento_docente: string;
    edad_docente: number;
    direccion_docente: string;
    telefono_docente: string;
    email_docente: string;
    titulo_docente: string;
    nivelEducacion_docente: string;
  }
  