// src/app/modules/docente/models/contratoDocenteModel.ts
export interface ContratoDocenteModel{
    id_contrato: number;
    docente_contrato: number;
    fecha_contrato: string;
    ciclo_contrato: string;
    nombramiento_contrato: string;
    especialidad_contrato: string;
    tipo_contrato: string;
    jornada_contrato: string;
    sueldo_contrato: number;
}

/**/