// src/app/modules/docente/models/docente.ts
export interface DocenteModel{
    id_docente: number;
    usuario_docente: number;
    nombres_docente: string;
    apellidos_docente: string;
    cedula_docente: string;
    fechaNacimiento_docente: string;
    edad_docente: number;
    direccion_docente: string;
    telefono_docente: string;
    email_docente: string;
    titulo_docente: string;
    nivelEducacion_docente: string;
    estado_docente: string;
}